+++ src/simd/avx2.cpp
@@ -52,7 +52,7 @@ float
 INT8L2Sqr(const void* pVect1v, const void* pVect2v, const void* qty_ptr) {
     auto* pVect1 = (int8_t*)pVect1v;
     auto* pVect2 = (int8_t*)pVect2v;
-    auto qty = *((size_t*)qty_ptr);
+    auto qty = *((uint64_t*)qty_ptr);
     return avx2::INT8ComputeL2Sqr(pVect1, pVect2, qty);
 }
 
@@ -912,7 +912,7 @@ PQFastScanLookUp32(const uint8_t* RESTRICT lookup_table,
         return;
     }
     __m256i sum[4];
-    for (size_t i = 0; i < 4; i++) {
+    for (uint64_t i = 0; i < 4; i++) {
         sum[i] = _mm256_setzero_si256();
     }
     const auto sign4 = _mm256_set1_epi8(0x0F);
@@ -1043,7 +1043,7 @@ BitNot(const uint8_t* x, const uint64_t num_byte, uint8_t* result) {
 }
 
 void
-VecRescale(float* data, size_t dim, float val) {
+VecRescale(float* data, uint64_t dim, float val) {
 #if defined(ENABLE_AVX2)
     int i = 0;
     __m256 val_vec = _mm256_set1_ps(val);
@@ -1076,10 +1076,10 @@ RotateOp(float* data, int idx, int dim_, int step) {
 }
 
 void
-FHTRotate(float* data, size_t dim_) {
+FHTRotate(float* data, uint64_t dim_) {
 #if defined(ENABLE_AVX2)
-    size_t n = dim_;
-    size_t step = 1;
+    uint64_t n = dim_;
+    uint64_t step = 1;
     while (step < n) {
         if (step >= 8) {
             avx2::RotateOp(data, 0, dim_, step);
@@ -1096,11 +1096,11 @@ FHTRotate(float* data, size_t dim_) {
 }
 
 void
-KacsWalk(float* data, size_t len) {
+KacsWalk(float* data, uint64_t len) {
 #if defined(ENABLE_AVX2)
-    size_t base = len % 2;
-    size_t offset = base + (len / 2);  // for odd dim
-    size_t i = 0;
+    uint64_t base = len % 2;
+    uint64_t offset = base + (len / 2);  // for odd dim
+    uint64_t i = 0;
     for (; i + 8 < len / 2; i += 8) {
         __m256 x = _mm256_loadu_ps(&data[i]);
         __m256 y = _mm256_loadu_ps(&data[i + offset]);
