+++ src/simd/pqfs_simd_test.cpp
@@ -29,7 +29,7 @@ compare_vector(std::vector<T>& v1, std::vector<T>& v2) {
     if (v1.size() != v2.size()) {
         return false;
     }
-    for (size_t i = 0; i < v1.size(); ++i) {
+    for (uint64_t i = 0; i < v1.size(); ++i) {
         if (v1[i] != v2[i]) {
             return false;
         }
